Fix #108: split default generator services
