+++ camel-kit-core/src/main/java/io/github/luigidemasi/camelkit/generator/OpenCodeGenerator.java
@@ -32,7 +32,7 @@ public class OpenCodeGenerator extends DefaultGenerator {
         Map<String, Object> data = new HashMap<>(
                 Map.of(
                         "COMMAND_PREFIX", ctx.commandPrefix()));
-        // AGENTS.md already generated by DefaultGenerator.generateAgentsMd()
+        // AGENTS.md already generated by the shared generator flow.
 
         // OpenCode-specific: generate agent definitions
         generateAgents(ctx);
