+++ camel-kit-core/src/main/java/io/github/luigidemasi/camelkit/generator/TemplateResourceCopier.java
@@ -1,0 +1,23 @@
+package io.github.luigidemasi.camelkit.generator;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+class TemplateResourceCopier {
+
+    void copy(String resourcePath, Path target) throws IOException {
+        try (InputStream is = getClass().getClassLoader().getResourceAsStream(resourcePath)) {
+            if (is == null) {
+                return;
+            }
+            Path parent = target.getParent();
+            if (parent != null) {
+                Files.createDirectories(parent);
+            }
+            Files.writeString(target, new String(is.readAllBytes(), StandardCharsets.UTF_8));
+        }
+    }
+}
