+++ camel-kit-core/src/main/java/io/github/luigidemasi/camelkit/generator/VersionPlaceholderResolver.java
@@ -1,0 +1,60 @@
+package io.github.luigidemasi.camelkit.generator;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Map;
+
+import io.github.luigidemasi.camelkit.CamelKitMain;
+import io.github.luigidemasi.camelkit.config.DistributionConfig;
+
+class VersionPlaceholderResolver {
+
+    void substitute(Path mdFile) throws IOException {
+        String content = Files.readString(mdFile);
+        Map<String, String> versionData = buildVersionTemplateData(CamelKitMain.distribution());
+
+        boolean hasPlaceholder = false;
+        for (String key : versionData.keySet()) {
+            if (content.contains("{" + key + "}")) {
+                hasPlaceholder = true;
+                break;
+            }
+        }
+        if (!hasPlaceholder) {
+            return;
+        }
+
+        String escaped = content.replace("{", "\\{");
+        for (String key : versionData.keySet()) {
+            escaped = escaped.replace("\\{" + key + "}", "{" + key + "}");
+        }
+
+        QuteTemplateEngine qute = new QuteTemplateEngine();
+        Map<String, Object> data = new java.util.HashMap<>(versionData);
+        String rendered = qute.renderString(escaped, data);
+        Files.writeString(mdFile, rendered);
+    }
+
+    private static Map<String, String> buildVersionTemplateData(DistributionConfig dist) {
+        Map<String, String> data = new java.util.LinkedHashMap<>();
+        data.put("CAMEL_VERSION", dist.camelMainVersion());
+        data.put("CAMEL_MAIN_VERSION", dist.camelMainVersion());
+        data.put("CAMEL_SPRINGBOOT_VERSION", dist.camelSpringbootVersion());
+        data.put("CAMEL_QUARKUS_VERSION", dist.camelQuarkusVersion());
+        data.put("SPRINGBOOT_BOM_VERSION", dist.springbootBomVersion());
+        data.put("QUARKUS_PLATFORM_VERSION", dist.quarkusPlatformVersion());
+        data.put("CAMEL_MAIN_SUPPORTED", dist.camelMainSupported());
+        data.put("CAMEL_SPRINGBOOT_SUPPORTED", dist.camelSpringbootSupported());
+        data.put("CAMEL_QUARKUS_SUPPORTED", dist.camelQuarkusSupported());
+
+        StringBuilder table = new StringBuilder();
+        for (var entry : dist.quarkusPlatformMappings().entrySet()) {
+            table.append("| ").append(entry.getKey())
+                    .append(" | ").append(entry.getValue())
+                    .append(" |\n");
+        }
+        data.put("QUARKUS_PLATFORM_TABLE", table.toString().stripTrailing());
+        return data;
+    }
+}
