+++ camel-kit-core/src/test/java/io/github/luigidemasi/camelkit/generator/DefaultGeneratorTest.java
@@ -229,35 +229,6 @@ class DefaultGeneratorTest {
     }
 
     @Test
-    void substitutionPreservesNonPlaceholderBraces() throws IOException {
-        Path mdFile = tempDir.resolve("test.md");
-        Files.writeString(mdFile, """
-                jackson: {}
-                ${quarkus.platform.version}
-                {COMMAND_PREFIX} graph stats
-                Version: {QUARKUS_PLATFORM_VERSION}
-                """);
-        DefaultGenerator.substituteVersionPlaceholders(mdFile);
-        String result = Files.readString(mdFile);
-
-        DistributionConfig dist = DistributionConfig.loadFromClasspathOrDefaults();
-        assertTrue(result.contains("jackson: {}"), "YAML empty map must be preserved");
-        assertTrue(result.contains("${quarkus.platform.version}"), "Maven property must be preserved");
-        assertTrue(result.contains("{COMMAND_PREFIX} graph stats"), "Non-version placeholder must be preserved");
-        assertTrue(result.contains(dist.quarkusPlatformVersion()), "Version placeholder must be resolved");
-        assertFalse(result.contains("{QUARKUS_PLATFORM_VERSION}"), "Version placeholder must not remain");
-    }
-
-    @Test
-    void substitutionSkipsFilesWithoutPlaceholders() throws IOException {
-        Path mdFile = tempDir.resolve("plain.md");
-        String original = "No placeholders here. Just some text with {braces} and ${maven}.";
-        Files.writeString(mdFile, original);
-        DefaultGenerator.substituteVersionPlaceholders(mdFile);
-        assertEquals(original, Files.readString(mdFile), "File without version placeholders must not change");
-    }
-
-    @Test
     void wrapsTomlForGemini() throws Exception {
         InitContext ctx = createContext("gemini");
         new DefaultGenerator().generate(ctx);
